Added a Lingscope AE for negation cue and scope detection.
It detects negation cue words with the baseline model which uses
a dictionary. It detects the scopes of the negations using a CRF
model that does not replace the cue word with the "CUE" string.
The Respective paper made these recommendations when you are
primarily interested into the scope of the negation. The baseline
cue model has a better coverage than the CRF cue detection model and
the replacement of the cue word with "CUE" did rather hurt then
help performance as reported by the paper(https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3000754/).